Separate Sequence and Timepoint implementations
Separate implementations of the Sequence and Timepoint classes
into their own files, and add typing hints where appropriate.
